refactor: adding nominal tests, refactoring components
