+++ src/components/Spinner/components/LoadingMessage.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+import { FONT_SIZE_XS } from 'style/styleVariables';
+
+export const LoadingMessage = styled.p`
+  opacity: 0.7;
+  font-size: ${FONT_SIZE_XS};
+  max-width: 50%;
+  margin: 0;
+`;
+
+export default LoadingMessage;
