+++ src/components/Spinner/components/LoadingMessage.test.js
@@ -1,0 +1,11 @@
+import React from "react";
+ import { shallow } from "enzyme";
+
+ import LoadingMessage from "./LoadingMessage";
+
+ describe("LoadingMessage", () => {
+   it("matches snapshot", () => {
+     const aLoadingMessage = shallow(<LoadingMessage/>);
+     expect(aLoadingMessage).toMatchSnapshot();
+   });
+ });
