+++ src/components/Spinner/components/LoadingSpinner.test.js
@@ -1,0 +1,11 @@
+import React from "react";
+ import { shallow } from "enzyme";
+
+ import LoadingSpinner from "./LoadingSpinner";
+
+ describe("LoadingSpinner", () => {
+   it("matches snapshot", () => {
+     const aLoadingSpinner = shallow(<LoadingSpinner/>);
+     expect(aLoadingSpinner).toMatchSnapshot();
+   });
+ });
